pyproject.toml
[tool.poetry]
name = "bridging"
version = "2.0.0"
description = "Bridging anaphora resolver for Japanese text"
license = "MIT"
authors = ["Nobuhiro Ueda <ueda@nlp.ist.i.kyoto-u.ac.jp>"]
maintainers = ["Nobuhiro Ueda <ueda@nlp.ist.i.kyoto-u.ac.jp>"]
readme = "README.md"
repository = "https://github.com/nobu-g/bridging-resolution"
keywords = [
    "NLP"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
torch = "~1.11.0"
transformers = "~4.16.0"
tokenizers = "^0.11.4"
pytorch-lightning = "^1.3"
kyoto-reader = "^2.5"
torchmetrics = "^0.6"
dataclasses-json = "^0.5.6"
ordered-set = "^4.0"
pandas = "^1.3"
scipy = "^1.7"
rich = "^10.12"
wandb = "^0.12.10"
hydra-core = "^1.1"
sentencepiece = "^0.1.96"
dash = "^2.0"
scikit-learn = "^1.0"

[tool.poetry.dev-dependencies]
ipdb = "*"
pytest = "*"
flake8 = "*"
pudb = "^2022.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
